requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
backcall==0.1.0
decorator==4.4.2
Django==3.0.5
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.1
graphql-relay==2.0.1
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
